+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Support for hierarchical hashtags! Hello, #books/2021
 
+### Fixed
+
+- Bug where things would get in an infinite loop if the buffer failed to parse.
+
 ## [0.35.0] - 2021-01-01 Happy New Year!
 
 ### Added